Add ability to automatically fetch roots from CCADB, fixes #65
- Add a cron job that once a day fetches roots from CCADB that are
  included in root stores and saves them into per-log root files (only
  appending certs for each log shard, but never removing them). This is
  similar to the sunlight implementation.
- Add config flag (enabled by default) to specify if a log will attempt
  to load its roots from the CCADB roots file in Workers KV.
- Remove roots.default.pem and require per-environment roots files to be
  non-empty if they exist. These can be used to add additional accepted
  roots on top of the CCADB list, for instance, to allow test/staging
  CAs to submit log entries.
- Add more functionality to x509util::CertPool to support adding and
  checking inclusion of certs in an existing CertPool.

+++ crates/ct_worker/build.rs
@@ -32,7 +32,7 @@ fn main() {
     let conf = serde_json::from_str::<AppConfig>(config_contents).unwrap_or_else(|e| {
         panic!("failed to deserialize JSON config '{config_file}': {e}");
     });
-    for (name, params) in conf.logs {
+    for (name, params) in &conf.logs {
         // Chrome's CT policy (https://googlechrome.github.io/CertificateTransparency/log_policy.html) states:
         // "The certificate expiry ranges for CT Logs must be no longer than one calendar year and should be no shorter than six months."
         assert!(
@@ -58,20 +58,35 @@ fn main() {
         }
     }
 
-    // Get and validate roots. Use 'roots.default.pem' if no environment-specific roots file is found.
-    let mut roots_file: &str = &format!("roots.{env}.pem");
-    if !fs::exists(roots_file).expect("failed to check if file exists") {
-        roots_file = "roots.default.pem";
+    // Get and validate roots from an embedded roots file, which must exist but
+    // can be empty unless 'enable_ccadb_roots' is set to false. If
+    // 'enable_ccadb_roots' is set to true, the log shard will combine trusted
+    // roots from the embedded roots file and from a roots file loaded from
+    // Workers KV.
+    let roots_file: &str = &format!("roots.{env}.pem");
+    let roots_file_exists = fs::exists(roots_file).expect("failed to check if file exists");
+    // If 'enable_ccadb_roots' is
+    if roots_file_exists {
+        let roots =
+            Certificate::load_pem_chain(&fs::read(roots_file).expect("failed to read roots file"))
+                .expect("unable to decode certificates");
+        assert!(
+            !roots.is_empty(),
+            "roots file does not contain any certificates"
+        );
+    } else if conf.logs.values().any(|params| !params.enable_ccadb_roots) {
+        // If any log shards have 'enable_ccadb_roots' set to false, require a roots file.
+        panic!("{roots_file} must exist and contain at least one certificate if any logs have 'enable_ccadb_roots' set to false");
     }
-    let roots =
-        Certificate::load_pem_chain(&fs::read(roots_file).expect("failed to read roots file"))
-            .expect("unable to decode certificates");
-    assert!(roots.len() > 50, "Roots file has too few entries");
 
     // Copy to OUT_DIR.
     let out_dir = env::var("OUT_DIR").unwrap();
     fs::copy(config_file, format!("{out_dir}/config.json")).expect("failed to copy config file");
-    fs::copy(roots_file, format!("{out_dir}/roots.pem")).expect("failed to copy roots file");
+    if roots_file_exists {
+        fs::copy(roots_file, format!("{out_dir}/roots.pem")).expect("failed to copy roots file");
+    } else {
+        fs::write(format!("{out_dir}/roots.pem"), "").expect("failed to write empty roots file");
+    }
 
     println!("cargo::rerun-if-env-changed=DEPLOY_ENV");
     println!("cargo::rerun-if-changed=config.schema.json");
